Reorganize modules.
