+++ spectrometer.cabal
@@ -487,6 +487,9 @@ library
     Strategy.Node.Npm.PackageLockV3
     Strategy.Node.PackageJson
     Strategy.Node.Pnpm.PnpmLock
+    Strategy.Node.Pnpm.Types
+    Strategy.Node.Pnpm.V4_8
+    Strategy.Node.Pnpm.V9
     Strategy.Node.Pnpm.Workspace
     Strategy.Node.YarnV1.YarnLock
     Strategy.Node.YarnV2.Lockfile
